dependencies installed and initiated index configs
