+++ modulo3/Labebank/src/teste.ts
@@ -1,0 +1,17 @@
+export class Client{
+  constructor(
+    name:string, 
+    cpf:number, 
+    birthDate:string){} 
+}
+
+export class CreateAccount{
+
+}
+
+export class Account{
+  constructor(client:Client, balance:number)
+  {}
+  
+}
+
